Attempt at mobile format fixes
Tried to fix mobile formatting. Everything works well except I need to make the content scale better at smaller resolutions. If I try to not scale and force the width of the view then I don't know how to force it to max zoom out with that max width so the user might always go back to the page partly zoomed in. Maybe find fix in later versions, this is fine for now.